start to refactor avatar loading stuff
