+++ packages/engine/src/avatar/functions/createAvatar.ts
@@ -10,12 +10,10 @@ import { AvatarComponent } from '../components/AvatarComponent'
 import { AvatarControllerComponent } from '../components/AvatarControllerComponent'
 import { Object3DComponent } from '../../scene/components/Object3DComponent'
 import { VelocityComponent } from '../../physics/components/VelocityComponent'
-import { CollisionGroups, DefaultCollisionMask } from '../../physics/enums/CollisionGroups'
+import { CollisionGroups } from '../../physics/enums/CollisionGroups'
 import { ColliderComponent } from '../../physics/components/ColliderComponent'
 import { AvatarAnimationComponent } from '../components/AvatarAnimationComponent'
 import { RaycastComponent } from '../../physics/components/RaycastComponent'
-import { Network } from '../../networking/classes/Network'
-import { AnimationGraph } from '../animations/AnimationGraph'
 import { AnimationState } from '../animations/AnimationState'
 import { InteractorComponent } from '../../interaction/components/InteractorComponent'
 import { NameComponent } from '../../scene/components/NameComponent'
@@ -28,11 +26,6 @@ import { useWorld } from '../../ecs/functions/SystemHooks'
 import { CollisionComponent } from '../../physics/components/CollisionComponent'
 import { SpawnPoseComponent } from '../components/SpawnPoseComponent'
 import { AvatarAnimationGraph } from '../animations/AvatarAnimationGraph'
-import { AudioTagComponent } from '../../audio/components/AudioTagComponent'
-import { ShadowComponent } from '../../scene/components/ShadowComponent'
-import { LocalInputTagComponent } from '../../input/components/LocalInputTagComponent'
-import { FollowCameraComponent, FollowCameraDefaultValues } from '../../camera/components/FollowCameraComponent'
-import { PersistTagComponent } from '../../scene/components/PersistTagComponent'
 import { createQuaternionProxy, createVector3Proxy } from '../../common/proxies/three'
 import { ObjectLayers } from '../../scene/constants/ObjectLayers'
 
@@ -140,12 +133,6 @@ export const createAvatar = (spawnAction: typeof NetworkWorldAction.spawnAvatar.
       rotation: new Quaternion().copy(spawnAction.parameters.rotation)
     })
     createAvatarController(entity)
-
-    addComponent(entity, AudioTagComponent, {})
-    addComponent(entity, ShadowComponent, { receiveShadow: true, castShadow: true })
-    addComponent(world.localClientEntity, LocalInputTagComponent, {})
-    addComponent(world.localClientEntity, FollowCameraComponent, FollowCameraDefaultValues)
-    addComponent(world.localClientEntity, PersistTagComponent, {})
   }
   const shape = world.physics.createShape(
     new PhysX.PxCapsuleGeometry(avatarRadius, capsuleHeight / 2),
