+++ packages/engine/src/ecs/functions/ComponentFunctions.ts
@@ -109,6 +109,7 @@ export const addComponent = <T, S extends bitECS.ISchema>(
   if (typeof entity === 'undefined') {
     throw new Error('[addComponent]: entity is undefined')
   }
+  if (hasComponent(entity, component)) throw new Error('component already exists' + entity + component._name)
   bitECS.addComponent(world, component, entity)
   if ((component as any)._schema) {
     for (const [key] of Object.entries((component as any)._schema as any)) {
